requirements.txt





elasticsearch==7.16.3
# Core
Django==3.2.4
pytz==2021.3
gunicorn==20.1.0
newrelic==6.4.0.157
django-dotenv==1.4.2
django-3-jet==1.0.8
celery==5.2.3 
django-celery-beat==2.2.1 
supervisor==4.2.4
# Rest apis
djangorestframework==3.12.4
Markdown==3.3.4
drf-yasg==1.20.0
django-filter==2.4.0
django-cors-headers==3.7.0
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
easy-thumbnails==2.7.1
django-auto-prefetching==0.1.10


PyMySQL==1.0.2 

django-redis==5.2.0  

numpy==1.21.2